extract functions to middlewareembeddedimages
